pyproject.toml
[tool.poetry]
name = "api-data-explorer"
version = "0.1.0"
description = ""
authors = ["Piotr Stempkowski <p.stempkowski@futuremind.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
matplot = "^0.1.9"
httpx = "^0.27.0"
pillow = "^10.2.0"
openpyxl = "^3.1.2"
pandas = "^2.2.0"
djangorestframework = "^3.14.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
